requirements.txt
certifi==2020.12.5
chardet==4.0.0
et-xmlfile==1.0.1
idna==2.10
kaleido==0.1.0
numpy==1.20.1
openpyxl==3.0.7
pandas==1.2.3
Pillow==8.1.2
plotly==4.14.3
psutil==5.8.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
retrying==1.3.3
six==1.15.0
typeguard==2.11.1
urllib3==1.26.3
